Clean up the build process
Now the build command is `npm run build` and I configured `netlify dev`
to run `elm-live` and redirect to its port. Turns out Netlify has
a nice document about netlify dev, it's just not linked in the main
docs (only through the CLI for some reason):
https://cli.netlify.com/netlify-dev

It was super useful!