Another massive refactor to state management
All made easy with Elm. ALSO!!! This is quite remarkable: I did not
try to build the architecture of Richard's SPA example, but it sort of
naturally happened??? Like, I called it State instead of Model,
but `Page/Quiz.elm` and `Page/QuizList.elm` are very much like the
page files in the SPA example. The next step is to split the view
code and that's pretty much it!